+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ### 201503301950
 
-* Added fiddles/jquery/fiddle-0006-VideoPlayer ~ [js fiddle](http://jsfiddle.net/bradyhouse/mszj46cf/)
-* Added fiddles/extjs/fiddle-20150330-BufferedDataGrid ~ [sencha fiddle](https://fiddle.sencha.com/#fiddle/kh9)
+* Added **fiddles/jquery/fiddle-0006-VideoPlayer** ~ [js fiddle](http://jsfiddle.net/bradyhouse/mszj46cf/)
+* Added **fiddles/extjs/fiddle-20150330-BufferedDataGrid** ~ [sencha fiddle](https://fiddle.sencha.com/#fiddle/kh9)
 * Added Updated LICENSE
 * Added .gitattributes
 * Added .editorconfig
+
+### 201503311810
+
+* Added **fiddles/extjs/fiddle-20150331-ExtJS4ColumnFiltering** ~ [sencha fiddle](https://fiddle.sencha.com/#fiddle/kie)
+* Added **fiddles/jquery/fiddle/fiddle-0007-BootstrapVideoCarousel** ~ [js fiddle](http://jsfiddle.net/bradyhouse/86mp2t4q/)